Feat: 기존의 자동차 객체 로직에서 자동차 이름 길이 예외 처리 로직 추가
