+++ README.md
@@ -1,4 +1,4 @@
-# 🚗 자동차 경주 게임 1,2단계 (Racing Car Game)
+# 🚗 자동차 경주 게임 1,2단계 (Racing Model.Car Game)
 
 주어진 횟수 동안 n대의 자동차가 무작위 값에 따라 전진 또는 멈추며, 최종적으로 가장 멀리 이동한 우승자를 가려내는 게임입니다.
 
@@ -11,7 +11,7 @@
 
 ## 🛠️ 기능 구현 목록
 
-### 1. 자동차 (Car)
+### 1. 자동차 (Model.Car)
 - [x] 자동차는 이름과 현재 위치 상태를 가진다.
 - [x] 무작위 값이 4 이상일 경우 위치를 1칸 전진한다.
 - [x] 무작위 값이 3 이하일 경우 현재 위치를 유지한다.
