+++ src/main/java/RacingGame.java
@@ -1,3 +1,5 @@
+import Model.Car;
+
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Random;
