+++ src/test/java/CarTest.java
@@ -1,4 +1,4 @@
-import static org.junit.jupiter.api.Assertions.*;
+import Model.Car;
 import org.junit.jupiter.api.Test;
 import static org.assertj.core.api.Assertions.assertThat;
