+++ src/test/java/RacingGameTest.java
@@ -1,3 +1,4 @@
+import Model.Car;
 import org.junit.jupiter.api.Test;
 import java.util.Arrays;
 import java.util.List;
